add tests for scaffolding
